🚚 rename: image component 분리
