+++ src/pages/PostDetail/index.tsx
@@ -5,7 +5,7 @@ import NoResults from "@/components/pages/search/NoResults";
 import UserCard from "@/components/pages/user/userList/UserCard";
 import LoadingSpinner from "@/components/ui/LoadingSpinner";
 import Page from "@/components/ui/Page";
-import PlaceholderImage from "@/components/ui/PlaceholderImage";
+
 import { removeImageFromStorage } from "@/lib/post/api";
 
 import { useAuthStore } from "@/stores/auth/useAuthStore";
@@ -16,6 +16,7 @@ import { useDeletePost } from "@/lib/post/hooks/useDeletePost";
 import { useGetPostByPostId } from "@/lib/post/hooks/useGetPostByPostId";
 import { PostImage } from "@/lib/post/type";
 
+import PlaceholderImage from "@/components/pages/posts/Image/PlaceholderImage";
 import { formatCount } from "@/shared/utils/formatCount";
 import { formatTimestamp } from "@/shared/utils/formatTimestamp";
 import { Timestamp } from "firebase/firestore";
